.
This repository houses a Random Joke Generator, a fun and entertaining web application that generates random jokes to bring laughter and amusement to users. Developed using JavaScript, HTML, and CSS, this generator fetches jokes from a predefined collection or an external API and displays them dynamically on the webpage. Users can simply click a button to generate a new joke, offering endless opportunities for laughter and enjoyment. With its simple and intuitive interface, this generator is easy to use and suitable for all audiences. Whether you're looking for a quick pick-me-up or a lighthearted distraction, this Random Joke Generator is sure to bring smiles to your face. Expand your JavaScript skills and spread joy with this delightful and humorous project.